refacto: Create a `file_uri` module containing `FileUri`, `TemplateUri` and `MultiFilesUri`
