+++ mithril-aggregator/src/file_uploaders/mod.rs
@@ -7,7 +7,7 @@ pub mod url_sanitizer;
 
 pub use dumb_uploader::*;
 pub use gcp_uploader::GcpUploader;
-pub use interface::{FileUploader, FileUri /*FilesUploader, MultiFilesUri, TemplateUri*/};
+pub use interface::FileUploader;
 pub use local_snapshot_uploader::LocalSnapshotUploader;
 pub use local_uploader::LocalUploader;
