+++ mithril-common/src/entities/cardano_database.rs
@@ -1,6 +1,3 @@
-use std::collections::HashSet;
-
-use anyhow::anyhow;
 use semver::Version;
 use serde::{Deserialize, Serialize};
 use sha2::{Digest, Sha256};
@@ -9,8 +6,9 @@ use strum::EnumDiscriminants;
 use crate::{
     entities::{CardanoDbBeacon, CompressionAlgorithm},
     signable_builder::Artifact,
-    StdResult,
 };
+
+use super::MultiFilesUri;
 
 /// Cardano database snapshot.
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
@@ -69,37 +67,6 @@ impl CardanoDatabaseSnapshot {
         hasher.update(self.merkle_root.as_bytes());
 
         hex::encode(hasher.finalize())
-    }
-}
-
-/// [TemplateUri] represents an URI pattern used to build a file's location
-#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Serialize, Deserialize)]
-pub struct TemplateUri(pub String);
-
-/// [MultiFilesUri] represents a unique location uri for multiple files
-#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Serialize, Deserialize)]
-pub enum MultiFilesUri {
-    /// URI template representing several URI
-    Template(TemplateUri),
-}
-
-impl MultiFilesUri {
-    /// Extract a template from a list of URIs
-    pub fn extract_template_from_uris(
-        file_uris: Vec<String>,
-        extractor: impl Fn(&str) -> StdResult<Option<String>>,
-    ) -> StdResult<Option<TemplateUri>> {
-        let mut templates = HashSet::new();
-        for file_uri in file_uris {
-            let template_uri = extractor(&file_uri)?;
-            template_uri.map(|template| templates.insert(template));
-        }
-
-        if templates.len() > 1 {
-            return Err(anyhow!("Multiple templates found in the file URIs"));
-        }
-
-        Ok(templates.into_iter().next().map(TemplateUri))
     }
 }
 
@@ -269,50 +236,6 @@ mod tests {
                 }
                 .compute_hash()
             );
-        }
-    }
-
-    mod extract_template_from_uris {
-        use super::*;
-
-        #[test]
-        fn returns_template() {
-            let file_uris = vec![
-                "http://whatever/00001.tar.gz".to_string(),
-                "http://whatever/00002.tar.gz".to_string(),
-            ];
-            fn extractor_returning_same_uri(_file_uri: &str) -> StdResult<Option<String>> {
-                Ok(Some(
-                    "http://whatever/{immutable_file_number}.tar.gz".to_string(),
-                ))
-            }
-
-            let template =
-                MultiFilesUri::extract_template_from_uris(file_uris, extractor_returning_same_uri)
-                    .unwrap();
-
-            assert_eq!(
-                template,
-                Some(TemplateUri(
-                    "http://whatever/{immutable_file_number}.tar.gz".to_string()
-                ))
-            );
-        }
-
-        #[test]
-        fn returns_error_with_multiple_templates() {
-            let file_uris = vec![
-                "http://whatever/00001.tar.gz".to_string(),
-                "http://00002.tar.gz/whatever".to_string(),
-            ];
-            fn extractor_returning_different_uri(file_uri: &str) -> StdResult<Option<String>> {
-                Ok(Some(file_uri.to_string()))
-            }
-
-            MultiFilesUri::extract_template_from_uris(file_uris, extractor_returning_different_uri)
-                .expect_err(
-                    "Should return an error when multiple templates are found in the file URIs",
-                );
         }
     }
 }
